Allow radio dialog to work after interupt
